feature/FSTB-83: present alert from MainView
